requirements.txt
alpaca-py>=0.8.2
pandas>=2.1.0
numpy>=1.26.0
ta-lib>=0.4.28
python-telegram-bot>=20.3
python-dotenv>=1.0.0
requests>=2.31.0
pandas-datareader>=0.10.0
lxml>=4.9.3
pytz>=2023.3
aiohttp>=3.8.5
beautifulsoup4>=4.12.2  # Required for pandas html parsing
asyncio==3.4.3
setuptools>=69.0.2
wheel>=0.42.0
certifi>=2024.2.2
urllib3>=2.1.0
nltk>=3.8.0 